Added front-end routing to navigate pages. Moved routing off of Server.js
