+++ controllers/index.js
@@ -1,9 +1,9 @@
 const router = require('express').Router();
 
 const apiRoutes = require('./api');
-// Placeholder for front end routes const homeRoutes = require('./homeRoutes');
+const homeRoutes = require('./homeRoutes');
 
-// Placeholder for front end routes router.use('/', homeRoutes);
+router.use('/', homeRoutes);
 router.use('/api', apiRoutes);
 
 module.exports = router;
